Add types, and forbid pcm_f32le.
Adding the types is needed for the build to succeed; I currently have
tsc in a pretty picky mode.

As for pcm_f32le, that's adding a header that Terry's equipment can't
read (as does pcm_s24le).  I might see about post-processing those
through the wavefile code to rewrite the header.